// compute/nextjs/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly authorId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz@1']['output'];
    };
  };
};